Remove old runner modules
